+++ src/testinferred.jl
@@ -4,6 +4,8 @@ export @testinferred
 
 using InteractiveUtils: gen_call_with_extracted_types
 using Test: @test
+using ..Utilities: args_and_call, materialize_broadcasted, typesplit, infer_return_type
+using InteractiveUtils: gen_call_with_extracted_types
 
 """
     @testinferred [AllowedType] f(x)
@@ -23,30 +25,6 @@ macro testinferred(allow, ex)
     return _inferred(ex, __module__, allow)
 end
 
-# helper functions
-_args_and_call((args..., f)...; kwargs...) = (args, kwargs, f(args...; kwargs...))
-_materialize_broadcasted(f, args...) = Broadcast.materialize(Broadcast.broadcasted(f, args...))
-@static if isdefined(Base, :typesplit)
-    const typesplit = Base.typesplit
-else
-    Base.@pure function typesplit(@nospecialize(a), @nospecialize(b))
-        a <: b && return Base.Bottom
-        isa(a, Union) && return Union{typesplit(a.a, b), typesplit(a.b, b)}
-        return a
-    end
-end
-
-# ensure backwards compatibility - have to use `return_types` which can return a vector of types
-@static if isdefined(Base, :infer_return_type)
-    infer_return_type(args...) = Base.infer_return_type(args...)
-else
-    function infer_return_type(args...)
-        inftypes = Base.return_types(args...)
-        @assert length(inftypes) == 1
-        return only(inftypes)
-    end
-end
-
 function _inferred(ex, mod, allow = :(Union{}))
     if Meta.isexpr(ex, :ref)
         ex = Expr(:call, :getindex, ex.args...)
@@ -57,7 +35,7 @@ function _inferred(ex, mod, allow = :(Union{}))
     farg = ex.args[1]
     if isa(farg, Symbol) && farg !== :.. && first(string(farg)) == '.'
         farg = Symbol(string(farg)[2:end])
-        ex = Expr(:call, GlobalRef(@__MODULE__, :_materialize_broadcasted), farg, ex.args[2:end]...)
+        ex = Expr(:call, GlobalRef(@__MODULE__, :materialize_broadcasted), farg, ex.args[2:end]...)
     end
 
     result = let ex = ex
@@ -70,7 +48,7 @@ function _inferred(ex, mod, allow = :(Union{}))
                         # Create the call expression with escaped user expressions
                         call_expr = :($(esc(ex.args[1]))(args...; kwargs...))
                         quote
-                            args, kwargs, result = $(esc(Expr(:call, _args_and_call, ex.args[2:end]..., ex.args[1])))
+                            args, kwargs, result = $(esc(Expr(:call, args_and_call, ex.args[2:end]..., ex.args[1])))
                             # wrap in dummy hygienic-scope to work around scoping issues with `call_expr` already having `esc` on the necessary parts
                             inftype = $(Expr(:var"hygienic-scope", gen_call_with_extracted_types(mod, infer_return_type, call_expr), @__MODULE__))
                         end
